Create component for Message Actions
