+++ res/css/_components.scss
@@ -171,6 +171,7 @@
 @import "./views/rooms/_LinkPreviewWidget.scss";
 @import "./views/rooms/_MemberInfo.scss";
 @import "./views/rooms/_MemberList.scss";
+@import "./views/rooms/_MessageActions.scss";
 @import "./views/rooms/_MessageComposer.scss";
 @import "./views/rooms/_MessageComposerFormatBar.scss";
 @import "./views/rooms/_NotificationBadge.scss";
