+++ res/css/views/rooms/_MessageActions.scss
@@ -1,0 +1,62 @@
+/*
+Copyright 2015, 2016 OpenMarket Ltd
+Copyright 2018 New Vector Ltd
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+.mx_MessageComposer_button {
+    position: relative;
+    margin-right: 12px;
+    cursor: pointer;
+    height: 20px;
+    width: 20px;
+
+    &::before {
+        content: '';
+        position: absolute;
+
+        height: 20px;
+        width: 20px;
+        background-color: $composer-button-color;
+        mask-repeat: no-repeat;
+        mask-size: contain;
+        mask-position: center;
+    }
+}
+
+.mx_MessageComposer_upload::before {
+    mask-image: url('$(res)/img/feather-customised/paperclip.svg');
+}
+
+.mx_MessageComposer_hangup::before {
+    mask-image: url('$(res)/img/hangup.svg');
+}
+
+.mx_MessageComposer_voicecall::before {
+    mask-image: url('$(res)/img/feather-customised/phone.svg');
+}
+
+.mx_MessageComposer_videocall::before {
+    mask-image: url('$(res)/img/feather-customised/video.svg');
+}
+
+.mx_MessageComposer_emoji::before {
+    mask-image: url('$(res)/img/feather-customised/emoji3.custom.svg');
+}
+
+.mx_MessageComposer_stickers::before {
+    mask-image: url('$(res)/img/feather-customised/sticker.custom.svg');
+}
+
+
