+++ res/css/views/rooms/_MessageComposer.scss
@@ -178,50 +178,6 @@ limitations under the License.
     color: $accent-color;
 }
 
-.mx_MessageComposer_button {
-    position: relative;
-    margin-right: 12px;
-    cursor: pointer;
-    height: 20px;
-    width: 20px;
-
-    &::before {
-        content: '';
-        position: absolute;
-
-        height: 20px;
-        width: 20px;
-        background-color: $composer-button-color;
-        mask-repeat: no-repeat;
-        mask-size: contain;
-        mask-position: center;
-    }
-}
-
-.mx_MessageComposer_upload::before {
-    mask-image: url('$(res)/img/feather-customised/paperclip.svg');
-}
-
-.mx_MessageComposer_hangup::before {
-    mask-image: url('$(res)/img/hangup.svg');
-}
-
-.mx_MessageComposer_voicecall::before {
-    mask-image: url('$(res)/img/feather-customised/phone.svg');
-}
-
-.mx_MessageComposer_videocall::before {
-    mask-image: url('$(res)/img/feather-customised/video.svg');
-}
-
-.mx_MessageComposer_emoji::before {
-    mask-image: url('$(res)/img/feather-customised/emoji3.custom.svg');
-}
-
-.mx_MessageComposer_stickers::before {
-    mask-image: url('$(res)/img/feather-customised/sticker.custom.svg');
-}
-
 .mx_MessageComposer_formatting {
     cursor: pointer;
     margin: 0 11px;
